+++ samples-azure-functions/src/main/java/com/functions/entities/counters.http
@@ -6,7 +6,7 @@
 // mode=0 or mode=entity (default) - dispatch to "counter" entity (TaskEntity<Integer>)
 // mode=1 or mode=state - dispatch to "counter_state" entity (POJO state dispatch)
 // mode=2 or mode=static - dispatch to "counter_alt" entity (same CounterEntity, different function name)
-// mode=3 or mode=manual - dispatch to "counter_manual" entity (ITaskEntity)
+// mode=3 or mode=manual - dispatch to "counter_manual" entity (TaskEntity)
 //
 // "counter" and "counter_alt" are the same entities, however they use
 // two different functions to dispatch, and thus are different entities when
